Hos_bot/requirements.txt
Pillow==8.2.0	
aiohttp==3.7.4.post0	
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
discord==1.0.1
discord.py==1.7.1
google==3.0.0
google-api-core==1.26.3
google-api-python-client==1.12.8
google-auth==1.29.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
h11==0.12.0	
httpcore==0.12.3
httplib2==0.19.1
httpx==0.17.1
idna==2.10
mpmath==1.2.1
multidict==5.1.0
packaging==20.9
pip==21.0.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyfiglet==0.8.post1
pyparsing==2.4.7
python-dotenv==0.17.0
python-resize-image==1.1.19
pytz==2021.1
requests==2.24.0
rfc3986==1.4.0
rsa==4.7.2
setuptools==56.0.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.0.1
sympy==1.8
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.9
yarl==1.6.3
youtube-search-python==1.4.5